refactor(qoe-metrics-dashboard): overview and detail models
Differentiate between requests between consumption and metrics
