+++ qoe-metrics-dashboard/src/app/models/types/metrics/requests/metrics-details-request-params.type.ts
@@ -1,0 +1,6 @@
+import { IQoeReport, IReceptionReport } from '../responses/metrics-details-report.interface';
+
+/**
+ * The endpoint allows to filter based on the properties of the reception report and the QoE report
+ */
+export type TMetricsDetailsRequestParams = IReceptionReport | Pick<IQoeReport, 'reportTime' | 'recordingSessionId'>;
