--- qoe-metrics-dashboard/src/app/models/types/requests/metrics-details-request-params.type.ts
@@ -1,4 +1,0 @@
-/**
- * The endpoint allows to filter based on the properties of the reception report and the QoE report
- */
-export type TMetricsDetailsRequestParams = Pick<any, 'reportTime' | 'clientID' | 'recordingSessionId'>;
